+++ README.md
@@ -13,11 +13,8 @@ pip install cn2date
 ```python
 from cn2date import parse
 
-result = parse("2021-09-17")
-# 输出结果：('2021-09-17 00:00:00', '2021-09-18 00:00:00')
-
-result = parse("2021年9月17日")
-# 输出结果：('2021-09-17 00:00:00', '2021-09-18 00:00:00')
+result = parse("2021年")
+# 输出结果：('2021-01-01 00:00:00', '2022-01-01 00:00:00')
 
 result = parse("二零二一年九月十七日")
 # 输出结果：('2021-09-17 00:00:00', '2021-09-18 00:00:00')
