+++ test/test_parse.py
@@ -1,448 +1,401 @@
-from pathlib import Path
-from test.util import get_node_value
+from datetime import datetime
+from typing import Tuple
 
 import pytest
-from lark import Lark
 
-file = Path(__file__).parent.parent / "cn2date/date.lark"
-file_text = open(file, "r", encoding="utf-8").read()
-grammars = file_text.split("===")
+from cn2date import parse
 
-date_parser = Lark(grammars[0])
-nl_parser = Lark(grammars[1])
+now = datetime.now()
+print("now", now)
 
+now_year = now.year
+now_month = now.month
+now_day = now.day
 
 date_testdata = [
     # 完整的日期格式
-    ("2021-9-10", "2021", "9", "10"),
-    ("2021-09-10", "2021", "09", "10"),
-    ("2021-09-1", "2021", "09", "1"),
-    ("2021-9-1", "2021", "9", "1"),
-    ("21-9-10", "21", "9", "10"),
-    ("21-09-10", "21", "09", "10"),
-    ("21-09-1", "21", "09", "1"),
-    ("21-9-1", "21", "9", "1"),
-    ("2021/9/10", "2021", "9", "10"),
-    ("2021/09/10", "2021", "09", "10"),
-    ("2021/09/1", "2021", "09", "1"),
-    ("2021/9/1", "2021", "9", "1"),
-    ("21/9/10", "21", "9", "10"),
-    ("21/09/10", "21", "09", "10"),
-    ("21/09/1", "21", "09", "1"),
-    ("21/9/1", "21", "9", "1"),
-    ("2021年9月10日", "2021", "9", "10"),
-    ("2021年09月10日", "2021", "09", "10"),
-    ("2021年09月10", "2021", "09", "10"),
-    ("2021年09月1", "2021", "09", "1"),
-    ("2021年9月1", "2021", "9", "1"),
-    ("21年9月10日", "21", "9", "10"),
-    ("21年09月10日", "21", "09", "10"),
-    ("21年09月10", "21", "09", "10"),
-    ("21年09月1", "21", "09", "1"),
-    ("21年9月1", "21", "9", "1"),
-    ("二零二一/九/三十一", "二零二一", "九", "三十一"),
-    ("二零二一/九/十", "二零二一", "九", "十"),
-    ("二零二一/零九/十", "二零二一", "零九", "十"),
-    ("二零二一/零九/一", "二零二一", "零九", "一"),
-    ("二零二一/九/一", "二零二一", "九", "一"),
-    ("二一/九/十", "二一", "九", "十"),
-    ("二一/零九/十", "二一", "零九", "十"),
-    ("二一/零九/一", "二一", "零九", "一"),
-    ("二一/九/一", "二一", "九", "一"),
-    ("二零二一-九-三十一", "二零二一", "九", "三十一"),
-    ("二零二一-九-十", "二零二一", "九", "十"),
-    ("二零二一-零九-十", "二零二一", "零九", "十"),
-    ("二零二一-零九-一", "二零二一", "零九", "一"),
-    ("二零二一-九-一", "二零二一", "九", "一"),
-    ("二一-九-十", "二一", "九", "十"),
-    ("二一-零九-十", "二一", "零九", "十"),
-    ("二一-零九-一", "二一", "零九", "一"),
-    ("二一-九-一", "二一", "九", "一"),
-    ("二零二一年九月三十一日", "二零二一", "九", "三十一"),
-    ("二零二一年九月三十一号", "二零二一", "九", "三十一"),
-    ("二零二一年九月三十一", "二零二一", "九", "三十一"),
-    ("二零二一年九月十日", "二零二一", "九", "十"),
-    ("二零二一年零九月十日", "二零二一", "零九", "十"),
-    ("二零二一年零九月十", "二零二一", "零九", "十"),
-    ("二零二一年零九月一", "二零二一", "零九", "一"),
-    ("二零二一年九月一", "二零二一", "九", "一"),
-    ("二一年九月十日", "二一", "九", "十"),
-    ("二一年零九月十日", "二一", "零九", "十"),
-    ("二一年零九月十", "二一", "零九", "十"),
-    ("二一年零九月一", "二一", "零九", "一"),
-    ("二一年九月一", "二一", "九", "一"),
-    ("二零二一年09月10号", "二零二一", "09", "10"),
-    ("二零二一年09/10号", "二零二一", "09", "10"),
-    ("二零二一/09月十号", "二零二一", "09", "十"),
-    ("二零二一年09-十日", "二零二一", "09", "十"),
-    ("二零二一年09-十号", "二零二一", "09", "十"),
-    ("二零二一年09-十", "二零二一", "09", "十"),
-    ("二零二一年09/10", "二零二一", "09", "10"),
-    ("二零二一年09-10", "二零二一", "09", "10"),
-    ("2021/九月10", "2021", "九", "10"),
-    ("2021/9月十", "2021", "9", "十"),
-    ("2021年9月10号", "2021", "9", "10"),
-    ("2021年09月10号", "2021", "09", "10"),
-    ("21年9月10号", "21", "9", "10"),
-    ("21年09月10号", "21", "09", "10"),
-    ("二零二一年九月十号", "二零二一", "九", "十"),
-    ("二零二一年零九月十号", "二零二一", "零九", "十"),
-    ("二一年九月十号", "二一", "九", "十"),
-    ("二一年零九月十号", "二一", "零九", "十"),
+    ("2017-7-23", (datetime(2017, 7, 23), datetime(2017, 7, 24))),
+    ("2017/7/23", (datetime(2017, 7, 23), datetime(2017, 7, 24))),
+    ("2017年7月23日", (datetime(2017, 7, 23), datetime(2017, 7, 24))),
+    ("二零一七年七月二十三", (datetime(2017, 7, 23), datetime(2017, 7, 24))),
+    ("二零一七年七月二十三日", (datetime(2017, 7, 23), datetime(2017, 7, 24))),
     # 只包含年的日期格式
-    ("21年", "21", None, None),
-    ("2021年", "2021", None, None),
-    ("二一年", "二一", None, None),
-    ("二零二一年", "二零二一", None, None),
+    ("17年", (datetime(2017, 1, 1), datetime(2018, 1, 1))),
+    ("2017年", (datetime(2017, 1, 1), datetime(2018, 1, 1))),
+    ("一七年", (datetime(2017, 1, 1), datetime(2018, 1, 1))),
+    ("二零一七年", (datetime(2017, 1, 1), datetime(2018, 1, 1))),
     # 只包含年月的日期格式
-    ("21-09", "21", "09", None),
-    ("21/09", "21", "09", None),
-    ("21年09月", "21", "09", None),
-    ("21年9月", "21", "9", None),
-    ("2021-09", "2021", "09", None),
-    ("2021/09", "2021", "09", None),
-    ("2021年09月", "2021", "09", None),
-    ("2021年9月", "2021", "9", None),
-    ("二一年零九月", "二一", "零九", None),
-    ("二一年九月", "二一", "九", None),
-    ("二零二一年零九月", "二零二一", "零九", None),
-    ("二零二一年九月", "二零二一", "九", None),
+    ("17-7", (datetime(2017, 7, 1), datetime(2017, 8, 1))),
+    ("17/7", (datetime(2017, 7, 1), datetime(2017, 8, 1))),
+    ("17年7月", (datetime(2017, 7, 1), datetime(2017, 8, 1))),
+    ("一七年七月", (datetime(2017, 7, 1), datetime(2017, 8, 1))),
+    ("2017-7", (datetime(2017, 7, 1), datetime(2017, 8, 1))),
+    ("2017/7", (datetime(2017, 7, 1), datetime(2017, 8, 1))),
+    ("2017年7月", (datetime(2017, 7, 1), datetime(2017, 8, 1))),
+    ("二零一七年七月", (datetime(2017, 7, 1), datetime(2017, 8, 1))),
     # 只包含月的日期格式
-    ("1月", None, "1", None),
-    ("01月", None, "01", None),
-    ("12月", None, "12", None),
-    ("一月", None, "一", None),
-    ("零一月", None, "零一", None),
-    ("十二月", None, "十二", None),
-    ("一二月", None, "一二", None),
+    ("7月", (datetime(now_year, 7, 1), datetime(now_year, 8, 1))),
+    ("07月", (datetime(now_year, 7, 1), datetime(now_year, 8, 1))),
+    ("七月", (datetime(now_year, 7, 1), datetime(now_year, 8, 1))),
     # 只包含月日的日期格式
-    ("9月10日", None, "9", "10"),
-    ("09月10日", None, "09", "10"),
-    ("09月1日", None, "09", "1"),
-    ("9月1日", None, "9", "1"),
-    ("9月10号", None, "9", "10"),
-    ("09月10号", None, "09", "10"),
-    ("09月1号", None, "09", "1"),
-    ("9月1号", None, "9", "1"),
-    ("九月十日", None, "九", "十"),
-    ("零九月十日", None, "零九", "十"),
-    ("零九月一日", None, "零九", "一"),
-    ("九月一日", None, "九", "一"),
-    ("九月十号", None, "九", "十"),
-    ("零九月十号", None, "零九", "十"),
-    ("零九月一号", None, "零九", "一"),
-    ("九月一号", None, "九", "一"),
+    # ("07-11", (datetime(now_year, 7, 11), datetime(now_year, 7, 12))), # 优先识别为 年-月
+    # ("07/11", (datetime(now_year, 7, 11), datetime(now_year, 7, 12))), # 优先识别为 年/月
+    ("07月11", (datetime(now_year, 7, 11), datetime(now_year, 7, 12))),
+    ("07月11日", (datetime(now_year, 7, 11), datetime(now_year, 7, 12))),
+    ("七月一一日", (datetime(now_year, 7, 11), datetime(now_year, 7, 12))),
+    ("七月十一日", (datetime(now_year, 7, 11), datetime(now_year, 7, 12))),
     # 只包含日的日期格式
-    ("1日", None, None, "1"),
-    ("01日", None, None, "01"),
-    ("31日", None, None, "31"),
-    ("一日", None, None, "一"),
-    ("零一日", None, None, "零一"),
-    ("三十一日", None, None, "三十一"),
-    ("三一日", None, None, "三一"),
+    ("7日", (datetime(now_year, now_month, 7), datetime(now_year, now_month, 8))),
+    ("07日", (datetime(now_year, now_month, 7), datetime(now_year, now_month, 8))),
+    ("七日", (datetime(now_year, now_month, 7), datetime(now_year, now_month, 8))),
+    ("7号", (datetime(now_year, now_month, 7), datetime(now_year, now_month, 8))),
+    ("07号", (datetime(now_year, now_month, 7), datetime(now_year, now_month, 8))),
+    ("七号", (datetime(now_year, now_month, 7), datetime(now_year, now_month, 8))),
 ]
 
 
 @pytest.mark.parametrize(
-    "text,expected_years,expected_months,expected_days",
-    date_testdata,
-    ids=[i[0] for i in date_testdata],
+    "text,expected", date_testdata, ids=[i[0] for i in date_testdata]
 )
-def test_date_parse(
-    text: str,
-    expected_years: str,
-    expected_months: str,
-    expected_days: str,
-):
+def test_date_parse(text: str, expected: Tuple[datetime, datetime]):
     print("input", text)
-    tree = date_parser.parse(text)
 
-    assert tree is not None
+    result = parse(text)
 
-    assert get_node_value(tree, "years") == expected_years
-    assert get_node_value(tree, "months") == expected_months
-    assert get_node_value(tree, "days") == expected_days
+    print("output", result)
+
+    assert len(result) == len(expected)
+
+    for i, s in enumerate(expected):
+        assert result[i] == s
 
 
 year_testdata = [
-    (nl_parser, "今年", None, None, None, None, "今年"),
-    (nl_parser, "本年", None, None, None, None, "本年"),
-    (nl_parser, "本年份", None, None, None, None, "本年"),
-    (nl_parser, "本年度", None, None, None, None, "本年"),
-    (nl_parser, "当前年", None, None, None, None, "当前年"),
-    (nl_parser, "当前年份", None, None, None, None, "当前年"),
-    (nl_parser, "当前年度", None, None, None, None, "当前年"),
-    (nl_parser, "明年", None, None, None, None, "明年"),
-    (nl_parser, "去年", None, None, None, None, "去年"),
-    (nl_parser, "前年", None, None, None, None, "前年"),
-    (nl_parser, "上半年", None, None, None, None, "上半年"),
-    (nl_parser, "下半年", None, None, None, None, "下半年"),
-    (nl_parser, "前两年", None, None, None, None, "前两年"),
-    (nl_parser, "前2年", None, None, None, None, "前2年"),
-    (nl_parser, "前两个年", None, None, None, None, "前两年"),
-    (nl_parser, "前2个年", None, None, None, None, "前2年"),
-    (nl_parser, "前两个年份", None, None, None, None, "前两年"),
-    (nl_parser, "前2个年份", None, None, None, None, "前2年"),
-    (nl_parser, "两年前", None, None, None, None, "两年前"),
-    (nl_parser, "2年前", None, None, None, None, "2年前"),
-    (nl_parser, "两年内", None, None, None, None, "两年内"),
-    (nl_parser, "2年内", None, None, None, None, "2年内"),
-    (nl_parser, "两年以前", None, None, None, None, "两年以前"),
-    (nl_parser, "2年以前", None, None, None, None, "2年以前"),
-    (nl_parser, "两年之前", None, None, None, None, "两年之前"),
-    (nl_parser, "2年之前", None, None, None, None, "2年之前"),
-    (nl_parser, "两年以内", None, None, None, None, "两年以内"),
-    (nl_parser, "2年以内", None, None, None, None, "2年以内"),
-    (nl_parser, "两年之内", None, None, None, None, "两年之内"),
-    (nl_parser, "2年之内", None, None, None, None, "2年之内"),
-    (nl_parser, "两年以来", None, None, None, None, "两年以来"),
-    (nl_parser, "2年以来", None, None, None, None, "2年以来"),
+    ("今年", (datetime(2021, 1, 1), datetime(2022, 1, 1))),
+    ("本年", (datetime(2021, 1, 1), datetime(2022, 1, 1))),
+    ("本年份", (datetime(2021, 1, 1), datetime(2022, 1, 1))),
+    ("本年度", (datetime(2021, 1, 1), datetime(2022, 1, 1))),
+    ("当前年", (datetime(2021, 1, 1), datetime(2022, 1, 1))),
+    ("当前年份", (datetime(2021, 1, 1), datetime(2022, 1, 1))),
+    ("当前年度", (datetime(2021, 1, 1), datetime(2022, 1, 1))),
+    ("明年", (datetime(2022, 1, 1), datetime(2023, 1, 1))),
+    ("去年", (datetime(2020, 1, 1), datetime(2021, 1, 1))),
+    ("前年", (datetime(2019, 1, 1), datetime(2020, 1, 1))),
+    ("今天", (datetime(2021, 9, 1), datetime(2021, 9, 2))),
+    ("上半年", (datetime(2021, 1, 1), datetime(2021, 7, 1))),
+    ("下半年", (datetime(2021, 7, 1), datetime(2022, 1, 1))),
+    ("前2年", (datetime(2019, 1, 1), datetime(2021, 1, 1))),
+    ("前2个年", (datetime(2019, 1, 1), datetime(2021, 1, 1))),
+    ("前2个年份", (datetime(2019, 1, 1), datetime(2021, 1, 1))),
+    ("前两年", (datetime(2019, 1, 1), datetime(2021, 1, 1))),
+    ("前两个年", (datetime(2019, 1, 1), datetime(2021, 1, 1))),
+    ("前两个年份", (datetime(2019, 1, 1), datetime(2021, 1, 1))),
+    ("后2年", (datetime(2022, 1, 1), datetime(2024, 1, 1))),
+    ("后2个年", (datetime(2022, 1, 1), datetime(2024, 1, 1))),
+    ("后2个年份", (datetime(2022, 1, 1), datetime(2024, 1, 1))),
+    ("后两年", (datetime(2022, 1, 1), datetime(2024, 1, 1))),
+    ("后两个年", (datetime(2022, 1, 1), datetime(2024, 1, 1))),
+    ("后两个年份", (datetime(2022, 1, 1), datetime(2024, 1, 1))),
+    ("2年前", (datetime(2019, 1, 1), datetime(2020, 1, 1))),
+    ("两年前", (datetime(2019, 1, 1), datetime(2020, 1, 1))),
+    ("2年后", (datetime(2023, 1, 1), datetime(2024, 1, 1))),
+    ("两年后", (datetime(2023, 1, 1), datetime(2024, 1, 1))),
+    ("2年内", (datetime(2020, 1, 1), datetime(2022, 1, 1))),
+    ("两年内", (datetime(2020, 1, 1), datetime(2022, 1, 1))),
 ]
 
 quarter_testdata = [
-    (nl_parser, "本季度", None, None, None, None, "本季度"),
-    (nl_parser, "这个季度", None, None, None, None, "这个季度"),
-    (nl_parser, "当前季度", None, None, None, None, "当前季度"),
-    (nl_parser, "上季度", None, None, None, None, "上季度"),
-    (nl_parser, "上个季度", None, None, None, None, "上季度"),
-    (nl_parser, "下季度", None, None, None, None, "下季度"),
-    (nl_parser, "下个季度", None, None, None, None, "下季度"),
-    (nl_parser, "第一季度", None, None, None, None, "一季度"),
-    (nl_parser, "第一个季度", None, None, None, None, "一季度"),
-    (nl_parser, "第1季度", None, None, None, None, "1季度"),
-    (nl_parser, "第1个季度", None, None, None, None, "1季度"),
-    (nl_parser, "一季度", None, None, None, None, "一季度"),
-    (nl_parser, "1季度", None, None, None, None, "1季度"),
-    (nl_parser, "第二季度", None, None, None, None, "二季度"),
-    (nl_parser, "第二个季度", None, None, None, None, "二季度"),
-    (nl_parser, "第2季度", None, None, None, None, "2季度"),
-    (nl_parser, "第2个季度", None, None, None, None, "2季度"),
-    (nl_parser, "2季度", None, None, None, None, "2季度"),
-    (nl_parser, "二季度", None, None, None, None, "二季度"),
-    (nl_parser, "第三季度", None, None, None, None, "三季度"),
-    (nl_parser, "第三个季度", None, None, None, None, "三季度"),
-    (nl_parser, "第3季度", None, None, None, None, "3季度"),
-    (nl_parser, "第3个季度", None, None, None, None, "3季度"),
-    (nl_parser, "三季度", None, None, None, None, "三季度"),
-    (nl_parser, "3季度", None, None, None, None, "3季度"),
-    (nl_parser, "第四季度", None, None, None, None, "四季度"),
-    (nl_parser, "第四个季度", None, None, None, None, "四季度"),
-    (nl_parser, "第4季度", None, None, None, None, "4季度"),
-    (nl_parser, "第4个季度", None, None, None, None, "4季度"),
-    (nl_parser, "四季度", None, None, None, None, "四季度"),
-    (nl_parser, "4季度", None, None, None, None, "4季度"),
-    (nl_parser, "前两季度", None, None, None, None, "前两季度"),
-    (nl_parser, "前2季度", None, None, None, None, "前2季度"),
-    (nl_parser, "前两个季度", None, None, None, None, "前两季度"),
-    (nl_parser, "前2个季度", None, None, None, None, "前2季度"),
-    (nl_parser, "两季度前", None, None, None, None, "两季度前"),
-    (nl_parser, "2季度前", None, None, None, None, "2季度前"),
-    (nl_parser, "两季度内", None, None, None, None, "两季度内"),
-    (nl_parser, "2季度内", None, None, None, None, "2季度内"),
-    (nl_parser, "两季度以来", None, None, None, None, "两季度以来"),
-    (nl_parser, "2季度以来", None, None, None, None, "2季度以来"),
-    (nl_parser, "两个季度前", None, None, None, None, "两季度前"),
-    (nl_parser, "2个季度前", None, None, None, None, "2季度前"),
-    (nl_parser, "两个季度以前", None, None, None, None, "两季度以前"),
-    (nl_parser, "2个季度以前", None, None, None, None, "2季度以前"),
-    (nl_parser, "两个季度之前", None, None, None, None, "两季度之前"),
-    (nl_parser, "2个季度之前", None, None, None, None, "2季度之前"),
-    (nl_parser, "两个季度以内", None, None, None, None, "两季度以内"),
-    (nl_parser, "2个季度以内", None, None, None, None, "2季度以内"),
-    (nl_parser, "两个季度之内", None, None, None, None, "两季度之内"),
-    (nl_parser, "2个季度之内", None, None, None, None, "2季度之内"),
-    (nl_parser, "两个季度以来", None, None, None, None, "两季度以来"),
-    (nl_parser, "2个季度以来", None, None, None, None, "2季度以来"),
+    ("本季度", (datetime(2021, 7, 1), datetime(2021, 10, 1))),
+    ("这个季度", (datetime(2021, 7, 1), datetime(2021, 10, 1))),
+    ("当前季度", (datetime(2021, 7, 1), datetime(2021, 10, 1))),
+    ("上季度", (datetime(2021, 4, 1), datetime(2021, 7, 1))),
+    ("上个季度", (datetime(2021, 4, 1), datetime(2021, 7, 1))),
+    ("下季度", (datetime(2021, 10, 1), datetime(2022, 1, 1))),
+    ("下个季度", (datetime(2021, 10, 1), datetime(2022, 1, 1))),
+    ("第一季度", (datetime(2021, 1, 1), datetime(2021, 4, 1))),
+    ("第一个季度", (datetime(2021, 1, 1), datetime(2021, 4, 1))),
+    ("第1季度", (datetime(2021, 1, 1), datetime(2021, 4, 1))),
+    ("第1个季度", (datetime(2021, 1, 1), datetime(2021, 4, 1))),
+    ("一季度", (datetime(2021, 1, 1), datetime(2021, 4, 1))),
+    ("1季度", (datetime(2021, 1, 1), datetime(2021, 4, 1))),
+    ("第二季度", (datetime(2021, 4, 1), datetime(2021, 7, 1))),
+    ("第二个季度", (datetime(2021, 4, 1), datetime(2021, 7, 1))),
+    ("第2季度", (datetime(2021, 4, 1), datetime(2021, 7, 1))),
+    ("第2个季度", (datetime(2021, 4, 1), datetime(2021, 7, 1))),
+    ("2季度", (datetime(2021, 4, 1), datetime(2021, 7, 1))),
+    ("二季度", (datetime(2021, 4, 1), datetime(2021, 7, 1))),
+    ("第三季度", (datetime(2021, 7, 1), datetime(2021, 10, 1))),
+    ("第三个季度", (datetime(2021, 7, 1), datetime(2021, 10, 1))),
+    ("第3季度", (datetime(2021, 7, 1), datetime(2021, 10, 1))),
+    ("第3个季度", (datetime(2021, 7, 1), datetime(2021, 10, 1))),
+    ("三季度", (datetime(2021, 7, 1), datetime(2021, 10, 1))),
+    ("3季度", (datetime(2021, 7, 1), datetime(2021, 10, 1))),
+    ("第四季度", (datetime(2021, 10, 1), datetime(2022, 1, 1))),
+    ("第四个季度", (datetime(2021, 10, 1), datetime(2022, 1, 1))),
+    ("第4季度", (datetime(2021, 10, 1), datetime(2022, 1, 1))),
+    ("第4个季度", (datetime(2021, 10, 1), datetime(2022, 1, 1))),
+    ("四季度", (datetime(2021, 10, 1), datetime(2022, 1, 1))),
+    ("4季度", (datetime(2021, 10, 1), datetime(2022, 1, 1))),
+    ("前2季度", (datetime(2021, 1, 1), datetime(2021, 7, 1))),
+    ("前2个季度", (datetime(2021, 1, 1), datetime(2021, 7, 1))),
+    ("前两季度", (datetime(2021, 1, 1), datetime(2021, 7, 1))),
+    ("前两个季度", (datetime(2021, 1, 1), datetime(2021, 7, 1))),
+    ("后2季度", (datetime(2021, 10, 1), datetime(2022, 4, 1))),
+    ("后2个季度", (datetime(2021, 10, 1), datetime(2022, 4, 1))),
+    ("后两季度", (datetime(2021, 10, 1), datetime(2022, 4, 1))),
+    ("后两个季度", (datetime(2021, 10, 1), datetime(2022, 4, 1))),
+    ("2季度前", (datetime(2021, 1, 1), datetime(2021, 4, 1))),
+    ("两季度前", (datetime(2021, 1, 1), datetime(2021, 4, 1))),
+    ("2季度后", (datetime(2022, 1, 1), datetime(2022, 4, 1))),
+    ("两季度后", (datetime(2022, 1, 1), datetime(2022, 4, 1))),
+    ("2季度内", (datetime(2021, 4, 1), datetime(2021, 10, 1))),
+    ("两季度内", (datetime(2021, 4, 1), datetime(2021, 10, 1))),
 ]
 
 month_testdata = [
-    (nl_parser, "本月", None, None, None, None, "本月"),
-    (nl_parser, "本月份", None, None, None, None, "本月"),
-    (nl_parser, "本月度", None, None, None, None, "本月"),
-    (nl_parser, "这个月", None, None, None, None, "这个月"),
-    (nl_parser, "这个月份", None, None, None, None, "这个月"),
-    (nl_parser, "当前月", None, None, None, None, "当前月"),
-    (nl_parser, "当前月份", None, None, None, None, "当前月"),
-    (nl_parser, "上月", None, None, None, None, "上月"),
-    (nl_parser, "上月份", None, None, None, None, "上月"),
-    (nl_parser, "上个月", None, None, None, None, "上月"),
-    (nl_parser, "上个月份", None, None, None, None, "上月"),
-    (nl_parser, "下月", None, None, None, None, "下月"),
-    (nl_parser, "下月份", None, None, None, None, "下月"),
-    (nl_parser, "下个月", None, None, None, None, "下月"),
-    (nl_parser, "下个月份", None, None, None, None, "下月"),
-    (nl_parser, "两月前", None, None, None, None, "两月前"),
-    (nl_parser, "2月前", None, None, None, None, "2月前"),
-    (nl_parser, "两月内", None, None, None, None, "两月内"),
-    (nl_parser, "2月内", None, None, None, None, "2月内"),
-    (nl_parser, "两月以前", None, None, None, None, "两月以前"),
-    (nl_parser, "两月之前", None, None, None, None, "两月之前"),
-    (nl_parser, "两月以内", None, None, None, None, "两月以内"),
-    (nl_parser, "2月以内", None, None, None, None, "2月以内"),
-    (nl_parser, "两月之内", None, None, None, None, "两月之内"),
-    (nl_parser, "2月之内", None, None, None, None, "2月之内"),
-    (nl_parser, "两月以来", None, None, None, None, "两月以来"),
-    (nl_parser, "两个月前", None, None, None, None, "两月前"),
-    (nl_parser, "2个月前", None, None, None, None, "2月前"),
-    (nl_parser, "两个月内", None, None, None, None, "两月内"),
-    (nl_parser, "2个月内", None, None, None, None, "2月内"),
-    (nl_parser, "两个月以前", None, None, None, None, "两月以前"),
-    (nl_parser, "2个月以前", None, None, None, None, "2月以前"),
-    (nl_parser, "两个月之前", None, None, None, None, "两月之前"),
-    (nl_parser, "2个月之前", None, None, None, None, "2月之前"),
-    (nl_parser, "两个月以内", None, None, None, None, "两月以内"),
-    (nl_parser, "2个月以内", None, None, None, None, "2月以内"),
-    (nl_parser, "两个月之内", None, None, None, None, "两月之内"),
-    (nl_parser, "2个月之内", None, None, None, None, "2月之内"),
-    (nl_parser, "两个月以来", None, None, None, None, "两月以来"),
-    (nl_parser, "2个月以来", None, None, None, None, "2月以来"),
-    (nl_parser, "前两月", None, None, None, None, "前两月"),
-    (nl_parser, "前2月", None, None, None, None, "前2月"),
-    (nl_parser, "前两个月", None, None, None, None, "前两月"),
-    (nl_parser, "前2个月", None, None, None, None, "前2月"),
+    ("本月", (datetime(2021, 9, 1), datetime(2021, 10, 1))),
+    ("本月份", (datetime(2021, 9, 1), datetime(2021, 10, 1))),
+    ("本月度", (datetime(2021, 9, 1), datetime(2021, 10, 1))),
+    ("这个月", (datetime(2021, 9, 1), datetime(2021, 10, 1))),
+    ("这个月份", (datetime(2021, 9, 1), datetime(2021, 10, 1))),
+    ("当前月", (datetime(2021, 9, 1), datetime(2021, 10, 1))),
+    ("当前月份", (datetime(2021, 9, 1), datetime(2021, 10, 1))),
+    ("上月", (datetime(2021, 8, 1), datetime(2021, 9, 1))),
+    ("上月份", (datetime(2021, 8, 1), datetime(2021, 9, 1))),
+    ("上个月", (datetime(2021, 8, 1), datetime(2021, 9, 1))),
+    ("上个月份", (datetime(2021, 8, 1), datetime(2021, 9, 1))),
+    ("下月", (datetime(2021, 10, 1), datetime(2021, 11, 1))),
+    ("下月份", (datetime(2021, 10, 1), datetime(2021, 11, 1))),
+    ("下个月", (datetime(2021, 10, 1), datetime(2021, 11, 1))),
+    ("下个月份", (datetime(2021, 10, 1), datetime(2021, 11, 1))),
+    ("前2月", (datetime(2021, 7, 1), datetime(2021, 9, 1))),
+    ("前2个月", (datetime(2021, 7, 1), datetime(2021, 9, 1))),
+    ("前2个月份", (datetime(2021, 7, 1), datetime(2021, 9, 1))),
+    ("前两月", (datetime(2021, 7, 1), datetime(2021, 9, 1))),
+    ("前两个月", (datetime(2021, 7, 1), datetime(2021, 9, 1))),
+    ("前两个月份", (datetime(2021, 7, 1), datetime(2021, 9, 1))),
+    ("后2月", (datetime(2021, 10, 1), datetime(2021, 12, 1))),
+    ("后2个月", (datetime(2021, 10, 1), datetime(2021, 12, 1))),
+    ("后2个月份", (datetime(2021, 10, 1), datetime(2021, 12, 1))),
+    ("后两月", (datetime(2021, 10, 1), datetime(2021, 12, 1))),
+    ("后两个月", (datetime(2021, 10, 1), datetime(2021, 12, 1))),
+    ("后两个月份", (datetime(2021, 10, 1), datetime(2021, 12, 1))),
+    ("2月前", (datetime(2021, 7, 1), datetime(2021, 8, 1))),
+    ("两月前", (datetime(2021, 7, 1), datetime(2021, 8, 1))),
+    ("2月后", (datetime(2021, 11, 1), datetime(2021, 12, 1))),
+    ("两月后", (datetime(2021, 11, 1), datetime(2021, 12, 1))),
+    ("2月内", (datetime(2021, 8, 1), datetime(2021, 10, 1))),
+    ("两月内", (datetime(2021, 8, 1), datetime(2021, 10, 1))),
 ]
 
 week_testdata = [
-    (nl_parser, "本周", None, None, None, None, "本周"),
-    (nl_parser, "当前周", None, None, None, None, "当前周"),
-    (nl_parser, "本星期", None, None, None, None, "本星期"),
-    (nl_parser, "这个星期", None, None, None, None, "这个星期"),
-    (nl_parser, "当前星期", None, None, None, None, "当前星期"),
-    (nl_parser, "上周", None, None, None, None, "上周"),
-    (nl_parser, "上星期", None, None, None, None, "上星期"),
-    (nl_parser, "上个星期", None, None, None, None, "上星期"),
-    (nl_parser, "下周", None, None, None, None, "下周"),
-    (nl_parser, "下星期", None, None, None, None, "下星期"),
-    (nl_parser, "下个星期", None, None, None, None, "下星期"),
-    (nl_parser, "前两星期", None, None, None, None, "前两星期"),
-    (nl_parser, "前2星期", None, None, None, None, "前2星期"),
-    (nl_parser, "前两个星期", None, None, None, None, "前两星期"),
-    (nl_parser, "前2个星期", None, None, None, None, "前2星期"),
-    (nl_parser, "两星期前", None, None, None, None, "两星期前"),
-    (nl_parser, "2星期前", None, None, None, None, "2星期前"),
-    (nl_parser, "两星期内", None, None, None, None, "两星期内"),
-    (nl_parser, "2星期内", None, None, None, None, "2星期内"),
-    (nl_parser, "两星期以来", None, None, None, None, "两星期以来"),
-    (nl_parser, "2星期以来", None, None, None, None, "2星期以来"),
-    (nl_parser, "两星期以内", None, None, None, None, "两星期以内"),
-    (nl_parser, "2星期以内", None, None, None, None, "2星期以内"),
-    (nl_parser, "两星期之内", None, None, None, None, "两星期之内"),
-    (nl_parser, "2星期之内", None, None, None, None, "2星期之内"),
-    (nl_parser, "两星期以来", None, None, None, None, "两星期以来"),
-    (nl_parser, "2星期以来", None, None, None, None, "2星期以来"),
-    (nl_parser, "两个星期前", None, None, None, None, "两星期前"),
-    (nl_parser, "2个星期前", None, None, None, None, "2星期前"),
-    (nl_parser, "两个星期以前", None, None, None, None, "两星期以前"),
-    (nl_parser, "2个星期以前", None, None, None, None, "2星期以前"),
-    (nl_parser, "两个星期之前", None, None, None, None, "两星期之前"),
-    (nl_parser, "2个星期之前", None, None, None, None, "2星期之前"),
-    (nl_parser, "两个星期内", None, None, None, None, "两星期内"),
-    (nl_parser, "2个星期内", None, None, None, None, "2星期内"),
-    (nl_parser, "两个星期以内", None, None, None, None, "两星期以内"),
-    (nl_parser, "2个星期以内", None, None, None, None, "2星期以内"),
-    (nl_parser, "两个星期之内", None, None, None, None, "两星期之内"),
-    (nl_parser, "2个星期之内", None, None, None, None, "2星期之内"),
-    (nl_parser, "两个星期以来", None, None, None, None, "两星期以来"),
-    (nl_parser, "2个星期以来", None, None, None, None, "2星期以来"),
+    ("本周", (datetime(2021, 8, 30), datetime(2021, 9, 6))),
+    ("当前周", (datetime(2021, 8, 30), datetime(2021, 9, 6))),
+    ("上周", (datetime(2021, 8, 23), datetime(2021, 8, 30))),
+    ("下周", (datetime(2021, 9, 6), datetime(2021, 9, 13))),
+    ("前2周", (datetime(2021, 8, 16), datetime(2021, 8, 30))),
+    ("前2个周", (datetime(2021, 8, 16), datetime(2021, 8, 30))),
+    ("前两周", (datetime(2021, 8, 16), datetime(2021, 8, 30))),
+    ("前两个周", (datetime(2021, 8, 16), datetime(2021, 8, 30))),
+    ("后2周", (datetime(2021, 9, 6), datetime(2021, 9, 20))),
+    ("后2个周", (datetime(2021, 9, 6), datetime(2021, 9, 20))),
+    ("后两周", (datetime(2021, 9, 6), datetime(2021, 9, 20))),
+    ("后两个周", (datetime(2021, 9, 6), datetime(2021, 9, 20))),
+    ("2周前", (datetime(2021, 8, 16), datetime(2021, 8, 23))),
+    ("两周前", (datetime(2021, 8, 16), datetime(2021, 8, 23))),
+    ("2周后", (datetime(2021, 9, 13), datetime(2021, 9, 20))),
+    ("两周后", (datetime(2021, 9, 13), datetime(2021, 9, 20))),
+    ("2周内", (datetime(2021, 8, 23), datetime(2021, 9, 6))),
+    ("两周内", (datetime(2021, 8, 23), datetime(2021, 9, 6))),
+    ("本星期", (datetime(2021, 8, 30), datetime(2021, 9, 6))),
+    ("这个星期", (datetime(2021, 8, 30), datetime(2021, 9, 6))),
+    ("当前星期", (datetime(2021, 8, 30), datetime(2021, 9, 6))),
+    ("上星期", (datetime(2021, 8, 23), datetime(2021, 8, 30))),
+    ("上个星期", (datetime(2021, 8, 23), datetime(2021, 8, 30))),
+    ("下星期", (datetime(2021, 9, 6), datetime(2021, 9, 13))),
+    ("下个星期", (datetime(2021, 9, 6), datetime(2021, 9, 13))),
+    ("前2星期", (datetime(2021, 8, 16), datetime(2021, 8, 30))),
+    ("前2个星期", (datetime(2021, 8, 16), datetime(2021, 8, 30))),
+    ("前两星期", (datetime(2021, 8, 16), datetime(2021, 8, 30))),
+    ("前两个星期", (datetime(2021, 8, 16), datetime(2021, 8, 30))),
+    ("后2星期", (datetime(2021, 9, 6), datetime(2021, 9, 20))),
+    ("后2个星期", (datetime(2021, 9, 6), datetime(2021, 9, 20))),
+    ("后两星期", (datetime(2021, 9, 6), datetime(2021, 9, 20))),
+    ("后两个星期", (datetime(2021, 9, 6), datetime(2021, 9, 20))),
+    ("2星期前", (datetime(2021, 8, 16), datetime(2021, 8, 23))),
+    ("两星期前", (datetime(2021, 8, 16), datetime(2021, 8, 23))),
+    ("2星期后", (datetime(2021, 9, 13), datetime(2021, 9, 20))),
+    ("两星期后", (datetime(2021, 9, 13), datetime(2021, 9, 20))),
+    ("2星期内", (datetime(2021, 8, 23), datetime(2021, 9, 6))),
+    ("两星期内", (datetime(2021, 8, 23), datetime(2021, 9, 6))),
 ]
 
 day_testdata = [
-    (nl_parser, "今天", None, None, None, None, "今天"),
-    (nl_parser, "今日", None, None, None, None, "今日"),
-    (nl_parser, "明天", None, None, None, None, "明天"),
-    (nl_parser, "明日", None, None, None, None, "明日"),
-    (nl_parser, "后天", None, None, None, None, "后天"),
-    (nl_parser, "后日", None, None, None, None, "后日"),
-    (nl_parser, "昨天", None, None, None, None, "昨天"),
-    (nl_parser, "昨日", None, None, None, None, "昨日"),
-    (nl_parser, "前天", None, None, None, None, "前天"),
-    (nl_parser, "前日", None, None, None, None, "前日"),
-    (nl_parser, "上午", None, None, None, None, "上午"),
-    (nl_parser, "下午", None, None, None, None, "下午"),
-    (nl_parser, "前两天", None, None, None, None, "前两天"),
-    (nl_parser, "前2天", None, None, None, None, "前2天"),
-    (nl_parser, "前两个天", None, None, None, None, "前两天"),
-    (nl_parser, "前2个天", None, None, None, None, "前2天"),
-    (nl_parser, "两天前", None, None, None, None, "两天前"),
-    (nl_parser, "2天前", None, None, None, None, "2天前"),
-    (nl_parser, "两天内", None, None, None, None, "两天内"),
-    (nl_parser, "2天内", None, None, None, None, "2天内"),
-    (nl_parser, "两天以来", None, None, None, None, "两天以来"),
-    (nl_parser, "2天以来", None, None, None, None, "2天以来"),
-    (nl_parser, "两天以内", None, None, None, None, "两天以内"),
-    (nl_parser, "2天以内", None, None, None, None, "2天以内"),
-    (nl_parser, "两天之内", None, None, None, None, "两天之内"),
-    (nl_parser, "2天之内", None, None, None, None, "2天之内"),
-    (nl_parser, "两天以来", None, None, None, None, "两天以来"),
-    (nl_parser, "2天以来", None, None, None, None, "2天以来"),
-    (nl_parser, "两个天前", None, None, None, None, "两天前"),
-    (nl_parser, "2个天前", None, None, None, None, "2天前"),
-    (nl_parser, "两个天以前", None, None, None, None, "两天以前"),
-    (nl_parser, "2个天以前", None, None, None, None, "2天以前"),
-    (nl_parser, "两个天之前", None, None, None, None, "两天之前"),
-    (nl_parser, "2个天之前", None, None, None, None, "2天之前"),
-    (nl_parser, "两个天内", None, None, None, None, "两天内"),
-    (nl_parser, "2个天内", None, None, None, None, "2天内"),
-    (nl_parser, "两个天以内", None, None, None, None, "两天以内"),
-    (nl_parser, "2个天以内", None, None, None, None, "2天以内"),
-    (nl_parser, "两个天之内", None, None, None, None, "两天之内"),
-    (nl_parser, "2个天之内", None, None, None, None, "2天之内"),
-    (nl_parser, "两个天以来", None, None, None, None, "两天以来"),
-    (nl_parser, "2个天以来", None, None, None, None, "2天以来"),
-    (nl_parser, "前两日", None, None, None, None, "前两日"),
-    (nl_parser, "前2日", None, None, None, None, "前2日"),
-    (nl_parser, "前两个日", None, None, None, None, "前两日"),
-    (nl_parser, "前2个日", None, None, None, None, "前2日"),
-    (nl_parser, "两日前", None, None, None, None, "两日前"),
-    (nl_parser, "2日前", None, None, None, None, "2日前"),
-    (nl_parser, "两日内", None, None, None, None, "两日内"),
-    (nl_parser, "2日内", None, None, None, None, "2日内"),
-    (nl_parser, "两日以来", None, None, None, None, "两日以来"),
-    (nl_parser, "两日以内", None, None, None, None, "两日以内"),
-    (nl_parser, "2日以内", None, None, None, None, "2日以内"),
-    (nl_parser, "两日之内", None, None, None, None, "两日之内"),
-    (nl_parser, "2日之内", None, None, None, None, "2日之内"),
-    (nl_parser, "两日以来", None, None, None, None, "两日以来"),
-    (nl_parser, "两个日前", None, None, None, None, "两日前"),
-    (nl_parser, "2个日前", None, None, None, None, "2日前"),
-    (nl_parser, "两个日以前", None, None, None, None, "两日以前"),
-    (nl_parser, "2个日以前", None, None, None, None, "2日以前"),
-    (nl_parser, "两个日之前", None, None, None, None, "两日之前"),
-    (nl_parser, "2个日之前", None, None, None, None, "2日之前"),
-    (nl_parser, "两个日内", None, None, None, None, "两日内"),
-    (nl_parser, "2个日内", None, None, None, None, "2日内"),
-    (nl_parser, "两个日以内", None, None, None, None, "两日以内"),
-    (nl_parser, "2个日以内", None, None, None, None, "2日以内"),
-    (nl_parser, "两个日之内", None, None, None, None, "两日之内"),
-    (nl_parser, "2个日之内", None, None, None, None, "2日之内"),
-    (nl_parser, "两个日以来", None, None, None, None, "两日以来"),
-    (nl_parser, "2个日以来", None, None, None, None, "2日以来"),
+    ("上午", (datetime(2021, 9, 1), datetime(2021, 9, 1, 12))),
+    ("下午", (datetime(2021, 9, 1, 12), datetime(2021, 9, 1, 19))),
+    ("今日", (datetime(2021, 9, 1), datetime(2021, 9, 2))),
+    ("明日", (datetime(2021, 9, 2), datetime(2021, 9, 3))),
+    ("后日", (datetime(2021, 9, 3), datetime(2021, 9, 4))),
+    ("昨日", (datetime(2021, 8, 31), datetime(2021, 9, 1))),
+    ("前日", (datetime(2021, 8, 30), datetime(2021, 8, 31))),
+    ("前2日", (datetime(2021, 8, 30), datetime(2021, 9, 1))),
+    ("前2个日", (datetime(2021, 8, 30), datetime(2021, 9, 1))),
+    ("前两日", (datetime(2021, 8, 30), datetime(2021, 9, 1))),
+    ("前两个日", (datetime(2021, 8, 30), datetime(2021, 9, 1))),
+    ("后2日", (datetime(2021, 9, 2), datetime(2021, 9, 4))),
+    ("后2个日", (datetime(2021, 9, 2), datetime(2021, 9, 4))),
+    ("后两日", (datetime(2021, 9, 2), datetime(2021, 9, 4))),
+    ("后两个日", (datetime(2021, 9, 2), datetime(2021, 9, 4))),
+    ("2日前", (datetime(2021, 8, 30), datetime(2021, 8, 31))),
+    ("两日前", (datetime(2021, 8, 30), datetime(2021, 8, 31))),
+    ("2日后", (datetime(2021, 9, 3), datetime(2021, 9, 4))),
+    ("两日后", (datetime(2021, 9, 3), datetime(2021, 9, 4))),
+    ("2日内", (datetime(2021, 8, 31), datetime(2021, 9, 2))),
+    ("两日内", (datetime(2021, 8, 31), datetime(2021, 9, 2))),
+    ("明天", (datetime(2021, 9, 2), datetime(2021, 9, 3))),
+    ("后天", (datetime(2021, 9, 3), datetime(2021, 9, 4))),
+    ("昨天", (datetime(2021, 8, 31), datetime(2021, 9, 1))),
+    ("前天", (datetime(2021, 8, 30), datetime(2021, 8, 31))),
+    ("前2天", (datetime(2021, 8, 30), datetime(2021, 9, 1))),
+    ("前2个天", (datetime(2021, 8, 30), datetime(2021, 9, 1))),
+    ("前两天", (datetime(2021, 8, 30), datetime(2021, 9, 1))),
+    ("前两个天", (datetime(2021, 8, 30), datetime(2021, 9, 1))),
+    ("后2天", (datetime(2021, 9, 2), datetime(2021, 9, 4))),
+    ("后2个天", (datetime(2021, 9, 2), datetime(2021, 9, 4))),
+    ("后两天", (datetime(2021, 9, 2), datetime(2021, 9, 4))),
+    ("后两个天", (datetime(2021, 9, 2), datetime(2021, 9, 4))),
+    ("2天前", (datetime(2021, 8, 30), datetime(2021, 8, 31))),
+    ("两天前", (datetime(2021, 8, 30), datetime(2021, 8, 31))),
+    ("2天后", (datetime(2021, 9, 3), datetime(2021, 9, 4))),
+    ("两天后", (datetime(2021, 9, 3), datetime(2021, 9, 4))),
+    ("2天内", (datetime(2021, 8, 31), datetime(2021, 9, 2))),
+    ("两天内", (datetime(2021, 8, 31), datetime(2021, 9, 2))),
 ]
 
 groupdate_testdata = [
-    ("2019年以前", "2019", None, None, "以前", None),
-    ("2019年之前", "2019", None, None, "之前", None),
-    ("2019年以来", "2019", None, None, "以来", None),
-    ("6月以前", None, "6", None, "以前", None),
-    ("6月之前", None, "6", None, "之前", None),
-    ("6月以来", None, "6", None, "以来", None),
-    ("13日以前", None, None, "13", "以前", None),
-    ("13日之前", None, None, "13", "之前", None),
-    ("13日以来", None, None, "13", "以来", None),
-    ("5号以前", None, None, "5", "以前", None),
-    ("5号之前", None, None, "5", "之前", None),
-    ("5号以来", None, None, "5", "以来", None),
-    ("2021年上半年", "2021", None, None, "上半年", None),
-    ("2021年第一季度", "2021", None, None, "一季度", None),
-    ("2021年9月7日上午", "2021", "9", "7", "上午", None),
+    ("2019年以前", (datetime.min, datetime(2019, 1, 1))),
+    ("19年以前", (datetime.min, datetime(2019, 1, 1))),
+    ("二零一九年以前", (datetime.min, datetime(2019, 1, 1))),
+    ("一九年以前", (datetime.min, datetime(2019, 1, 1))),
+    ("2019年以后", (datetime(2020, 1, 1), datetime.max)),
+    ("19年以后", (datetime(2020, 1, 1), datetime.max)),
+    ("二零一九年以后", (datetime(2020, 1, 1), datetime.max)),
+    ("一九年以后", (datetime(2020, 1, 1), datetime.max)),
+    ("2019年以来", (datetime(2019, 1, 1), datetime(2021, 9, 2))),
+    ("19年以来", (datetime(2019, 1, 1), datetime(2021, 9, 2))),
+    ("二零一九年以来", (datetime(2019, 1, 1), datetime(2021, 9, 2))),
+    ("一九年以来", (datetime(2019, 1, 1), datetime(2021, 9, 2))),
+    ("2017年上半年", (datetime(2017, 1, 1), datetime(2017, 7, 1))),
+    ("17年上半年", (datetime(2017, 1, 1), datetime(2017, 7, 1))),
+    ("二零一七年上半年", (datetime(2017, 1, 1), datetime(2017, 7, 1))),
+    ("一七年上半年", (datetime(2017, 1, 1), datetime(2017, 7, 1))),
+    ("2017年下半年", (datetime(2017, 7, 1), datetime(2018, 1, 1))),
+    ("17年下半年", (datetime(2017, 7, 1), datetime(2018, 1, 1))),
+    ("二零一七年下半年", (datetime(2017, 7, 1), datetime(2018, 1, 1))),
+    ("一七年下半年", (datetime(2017, 7, 1), datetime(2018, 1, 1))),
+    ("2019年第一季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("2019年第一个季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("2019年一季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("2019年第1季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("2019年第1个季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("2019年1季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("19年第一季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("19年第一个季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("19年一季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("19年第1季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("19年第1个季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("19年1季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("二零一九年第一季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("二零一九年第一个季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("二零一九年一季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("二零一九年第1季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("二零一九年第1个季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("二零一九年1季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("一九年第一季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("一九年第一个季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("一九年一季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("一九年第1季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("一九年第1个季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("一九年1季度", (datetime(2019, 1, 1), datetime(2019, 4, 1))),
+    ("2019年第二季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("2019年第二个季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("2019年二季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("2019年第2季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("2019年第2个季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("2019年2季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("19年第二季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("19年第二个季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("19年二季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("19年第2季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("19年第2个季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("19年2季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("二零一九年第二季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("二零一九年第二个季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("二零一九年二季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("二零一九年第2季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("二零一九年第2个季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("二零一九年2季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("一九年第二季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("一九年第二个季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("一九年二季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("一九年第2季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("一九年第2个季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("一九年2季度", (datetime(2019, 4, 1), datetime(2019, 7, 1))),
+    ("2019年第三季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("2019年第三个季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("2019年三季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("2019年第3季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("2019年第3个季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("2019年3季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("19年第三季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("19年第三个季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("19年三季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("19年第3季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("19年第3个季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("19年3季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("二零一九年第三季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("二零一九年第三个季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("二零一九年三季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("二零一九年第3季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("二零一九年第3个季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("二零一九年3季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("一九年第三季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("一九年第三个季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("一九年三季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("一九年第3季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("一九年第3个季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("一九年3季度", (datetime(2019, 7, 1), datetime(2019, 10, 1))),
+    ("2019年第四季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("2019年第四个季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("2019年四季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("2019年第4季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("2019年第4个季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("2019年4季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("19年第四季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("19年第四个季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("19年四季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("19年第4季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("19年第4个季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("19年4季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("二零一九年第四季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("二零一九年第四个季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("二零一九年四季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("二零一九年第4季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("二零一九年第4个季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("二零一九年4季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("一九年第四季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("一九年第四个季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("一九年四季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("一九年第4季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("一九年第4个季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
+    ("一九年4季度", (datetime(2019, 10, 1), datetime(2020, 1, 1))),
 ]
 
 parse_testdata = [
-    *date_testdata,
-    # *year_testdata,
-    # *quarter_testdata,
-    # *month_testdata,
-    # *week_testdata,
-    # *day_testdata,
-    # *groupdate_testdata,
+    *year_testdata,
+    *quarter_testdata,
+    *month_testdata,
+    *week_testdata,
+    *day_testdata,
+    # *groupdate_testdata, # TODO: 待完善组合日期的逻辑
 ]
